# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rheo API",
    "description": "Rheo Sovereign AI Voice Studio & Neural Speech Synthesis Engine \u2014 Architected by masterz1311",
    "version": "0.5.0"
  },
  "paths": {
...
    "/models/load": {
      "post": {
        "summary": "Load Model",
        "description": "Manually load TTS model.",
        "operationId": "load_model_models_load_post",
        "parameters": [
          {
            "name": "model_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(1\\.7B|0\\.6B|1B|3B)$",
              "default": "1.7B",
              "title": "Model Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/models/{model_name}/unload": {
      "post": {
        "summary": "Unload Model By Name",
        "description": "Unload a specific model from memory without deleting it from disk.",
        "operationId": "unload_model_by_name_models__model_name__unload_post",
        "parameters": [
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/progress/{model_name}": {
      "get": {
        "summary": "Get Model Progress",
        "description": "Get model download progress via Server-Sent Events.",
        "operationId": "get_model_progress_models_progress__model_name__get",
        "parameters": [
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/models/migrate": {
      "post": {
        "summary": "Migrate Models",
        "description": "Move all downloaded models to a new directory with byte-level progress via SSE.",
        "operationId": "migrate_models_models_migrate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelMigrateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/models/download": {
      "post": {
        "summary": "Trigger Model Download",
        "description": "Trigger download of a specific model.",
        "operationId": "trigger_model_download_models_download_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/download/cancel": {
      "post": {
        "summary": "Cancel Model Download",
        "description": "Cancel or dismiss an errored/stale download task.",
        "operationId": "cancel_model_download_models_download_cancel_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{model_name}": {
      "delete": {
        "summary": "Delete Model",
        "description": "Delete a downloaded model from the HuggingFace cache.",
        "operationId": "delete_model_models__model_name__delete",
        "parameters": [
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelDownloadRequest": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name"
          }
        },
        "type": "object",
        "required": [
          "model_name"
        ],
        "title": "ModelDownloadRequest",
        "description": "Request model for triggering model download."
      },
      "ModelMigrateRequest": {
        "properties": {
          "destination": {
            "type": "string",
            "title": "Destination"
          }
        },
        "type": "object",
        "required": [
          "destination"
        ],
        "title": "ModelMigrateRequest",
        "description": "Request model for migrating models to a new directory."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}